GPU-842: Fixed possibility to perform actions on deleted items which …

+++ cms-ui/apps/editor-ui/src/app/list-view/components/item-list-header/item-list-header.component.ts
@@ -174,7 +174,7 @@ export class ItemListHeaderComponent implements OnInit, OnChanges, OnDestroy {
         this.subscriptions.push(this.appState.select(state => state.features.wastebin).subscribe(enabled => {
             this.wastebinEnabled = enabled;
             this.changeDetector.markForCheck();
-        }))
+        }));
     }
 
     ngOnChanges(changes: SimpleChanges): void {
@@ -308,29 +308,65 @@ export class ItemListHeaderComponent implements OnInit, OnChanges, OnDestroy {
     }
 
     /**
+     * Helper method to filter out deleted items and which shows a notification if
+     * no regular/not deleted item has been selected yet.
+     */
+    private getNotDeletedItems(): Item[] {
+        const validItems = this.selectedItems.filter(item => !EntityStateUtil.stateDeleted(item));
+        if (validItems.length !== 0) {
+            return validItems;
+        }
+
+        this.notifications.show({
+            message: 'editor.select_not_deleted_items',
+            translationParams: {
+                itemTypePlural: this.i18n.translate(`common.type_${this.itemType}s`),
+            },
+            type: 'warning',
+        });
+        return [];
+
+    }
+
+    /**
      * Copy the selected items to a different folder and clear the selection.
      */
     copySelected(): void {
-        this.contextMenuOperations.copyItems(this.itemType, this.selectedItems, this.activeNode.id);
+        const itemsToCopy = this.getNotDeletedItems();
+        if (itemsToCopy.length === 0) {
+            return;
+        }
+
+        this.contextMenuOperations.copyItems(this.itemType, itemsToCopy, this.activeNode.id);
     }
 
     /**
      * Move the selected items to a different folder and clear the selection.
      */
     moveSelected(): void {
-        this.contextMenuOperations.moveItems(this.itemType, this.selectedItems, this.activeNode.id, this.currentFolderId);
+        const itemsToMove = this.getNotDeletedItems();
+        if (itemsToMove.length === 0) {
+            return;
+        }
+
+        this.contextMenuOperations.moveItems(this.itemType, itemsToMove, this.activeNode.id, this.currentFolderId);
     }
 
     /**
      * Publish the selected pages and clear the selection.
      */
     publishSelected(type: FolderItemType): void {
+        const itemsToPublish = this.getNotDeletedItems();
+        if (itemsToPublish.length === 0) {
+            return;
+        }
+
         switch (type) {
             case 'page':
-                this.contextMenuOperations.publishPages(this.selectedItems as Page[]);
+                this.contextMenuOperations.publishPages(itemsToPublish as Page[]);
                 break;
             case 'form':
-                this.contextMenuOperations.publishForms(this.selectedItems as Form[]);
+                this.contextMenuOperations.publishForms(itemsToPublish as Form[]);
                 break;
             default:
                 throw new Error(`Behavior not defined for type: "${type}".`);
@@ -341,19 +377,29 @@ export class ItemListHeaderComponent implements OnInit, OnChanges, OnDestroy {
      * Publish the selected pages and clear the selection.
      */
     publishLanguageVariantsSelected(): void {
-        this.contextMenuOperations.publishPages(this.selectedItems as Page[], true);
+        const itemsToPublish = this.getNotDeletedItems();
+        if (itemsToPublish.length === 0) {
+            return;
+        }
+
+        this.contextMenuOperations.publishPages(itemsToPublish as Page[], true);
     }
 
     /**
      * Take the selected pages offline and clear the selection.
      */
     takeSelectedOffline(type: ItemType): void {
+        const itemsToTakeOffline = this.getNotDeletedItems();
+        if (itemsToTakeOffline.length === 0) {
+            return;
+        }
+
         switch (type) {
             case 'page':
-                this.contextMenuOperations.takePagesOffline(this.selectedItems as Page[]);
+                this.contextMenuOperations.takePagesOffline(itemsToTakeOffline as Page[]);
                 break;
             case 'form':
-                this.contextMenuOperations.takeFormsOffline(this.selectedItems as Form[]);
+                this.contextMenuOperations.takeFormsOffline(itemsToTakeOffline as Form[]);
                 break;
             default:
                 throw new Error(`Behavior not defined for type: "${type}".`);
@@ -364,14 +410,23 @@ export class ItemListHeaderComponent implements OnInit, OnChanges, OnDestroy {
      * Take the selected items offline and clear the selection.
      */
     localizeSelected(): void {
-        this.contextMenuOperations.localizeItems(this.itemType, this.selectedItems as InheritableItem[], this.activeNode.id);
+        const itemsToBeLocalized = this.getNotDeletedItems();
+        if (itemsToBeLocalized.length === 0) {
+            return;
+        }
+
+        this.contextMenuOperations.localizeItems(this.itemType, itemsToBeLocalized as InheritableItem[], this.activeNode.id);
     }
 
     /**
      * Delete the selected items and clear the selection.
      */
     deleteSelected(type: FolderItemType): void {
-        const itemsToBeDeleted = this.selectedItems.filter(item => !EntityStateUtil.stateDeleted(item));
+        const itemsToBeDeleted = this.getNotDeletedItems();
+        if (itemsToBeDeleted.length === 0) {
+            return;
+        }
+
         this.contextMenuOperations.deleteItems(type, itemsToBeDeleted, this.activeNode.id);
     }
 
@@ -379,8 +434,20 @@ export class ItemListHeaderComponent implements OnInit, OnChanges, OnDestroy {
      * Restore the selected items and clear the selection.
      */
     restoreSelected(): void {
-        const itemsToBeDeleted = this.selectedItems.filter(item => EntityStateUtil.stateDeleted(item));
-        this.contextMenuOperations.restoreItems(itemsToBeDeleted);
+        const itemsToBeRestored = this.selectedItems.filter(item => EntityStateUtil.stateDeleted(item));
+        if (itemsToBeRestored.length === 0) {
+            this.notifications.show({
+                message: 'editor.select_deleted_items',
+                translationParams: {
+                    itemTypePlural: this.i18n.translate(`common.type_${this.itemType}s`),
+                },
+                type: 'warning',
+            });
+
+            return;
+        }
+
+        this.contextMenuOperations.restoreItems(itemsToBeRestored);
     }
 
     /**
@@ -429,14 +496,24 @@ export class ItemListHeaderComponent implements OnInit, OnChanges, OnDestroy {
         this.userSettings.setSorting(type, sorting.sortBy, sorting.sortOrder);
     }
 
-    createVariationsClicked(pages: Page[]): void {
-        this.contextMenuOperations.createVariationsClicked(pages, this.activeNode.id);
+    createVariationsClicked(): void {
+        const pagesToCreateVariationsFrom = this.getNotDeletedItems();
+        if (pagesToCreateVariationsFrom.length === 0) {
+            return;
+        }
+
+        this.contextMenuOperations.createVariationsClicked(pagesToCreateVariationsFrom as Page[], this.activeNode.id);
     }
 
     async stageItems(mode: StagingMode): Promise<void> {
+        const itemsToStage = this.getNotDeletedItems();
+        if (itemsToStage.length === 0) {
+            return;
+        }
+
         let counter = 0;
 
-        for (const item of this.selectedItems) {
+        for (const item of itemsToStage) {
             let res: boolean;
             if (this.itemType === 'page') {
                 res = await this.contextMenuOperations.stagePageToCurrentPackage(item as Page, mode === StagingMode.ALL_LANGUAGES, false);
@@ -448,14 +525,14 @@ export class ItemListHeaderComponent implements OnInit, OnChanges, OnDestroy {
             }
         }
 
-        if (counter === this.selectedItems.length) {
+        if (counter === itemsToStage.length) {
             if (counter === 1) {
                 this.notifications.show({
                     type: 'success',
                     message: 'editor.stage_item_success_message',
                     translationParams: {
                         itemType: this.i18n.translate(`common.type_${this.itemType}_article`),
-                        itemName: this.selectedItems[0].name,
+                        itemName: itemsToStage[0].name,
                     },
                 });
             } else {
@@ -464,7 +541,7 @@ export class ItemListHeaderComponent implements OnInit, OnChanges, OnDestroy {
                     message: 'editor.stage_multiple_items_success_message',
                     translationParams: {
                         itemType: this.i18n.translate(`common.type_${plural[this.itemType]}`),
-                        amount: this.selectedItems.length,
+                        amount: itemsToStage.length,
                     },
                 });
             }
@@ -474,7 +551,7 @@ export class ItemListHeaderComponent implements OnInit, OnChanges, OnDestroy {
                 message: 'editor.stage_multiple_items_partial_success_message',
                 translationParams: {
                     itemType: this.i18n.translate(`common.type_${plural[this.itemType]}`),
-                    amount: this.selectedItems.length,
+                    amount: itemsToStage.length,
                     count: counter,
                 },
             });
@@ -483,9 +560,14 @@ export class ItemListHeaderComponent implements OnInit, OnChanges, OnDestroy {
     }
 
     async unstageItems(mode: StagingMode): Promise<void> {
+        const itemsToUnstage = this.getNotDeletedItems();
+        if (itemsToUnstage.length === 0) {
+            return;
+        }
+
         let counter = 0;
 
-        for (const item of this.selectedItems) {
+        for (const item of itemsToUnstage) {
             let res: boolean;
             if (this.itemType === 'page') {
                 res = await this.contextMenuOperations.unstagePageFromCurrentPackage(item as Page, mode === StagingMode.ALL_LANGUAGES, false);
@@ -497,14 +579,14 @@ export class ItemListHeaderComponent implements OnInit, OnChanges, OnDestroy {
             }
         }
 
-        if (counter === this.selectedItems.length) {
+        if (counter === itemsToUnstage.length) {
             if (counter === 1) {
                 this.notifications.show({
                     type: 'success',
                     message: 'editor.unstage_item_success_message',
                     translationParams: {
                         itemType: this.i18n.translate(`common.type_${this.itemType}_article`),
-                        itemName: this.selectedItems[0].name,
+                        itemName: itemsToUnstage[0].name,
                     },
                 });
             } else {
@@ -513,7 +595,7 @@ export class ItemListHeaderComponent implements OnInit, OnChanges, OnDestroy {
                     message: 'editor.unstage_multiple_items_success_message',
                     translationParams: {
                         itemType: this.i18n.translate(`common.type_${plural[this.itemType]}`),
-                        amount: this.selectedItems.length,
+                        amount: itemsToUnstage.length,
                     },
                 });
             }
@@ -523,7 +605,7 @@ export class ItemListHeaderComponent implements OnInit, OnChanges, OnDestroy {
                 message: 'editor.unstage_multiple_items_partial_success_message',
                 translationParams: {
                     itemType: this.i18n.translate(`common.type_${plural[this.itemType]}`),
-                    amount: this.selectedItems.length,
+                    amount: itemsToUnstage.length,
                     count: counter,
                 },
             });
